* Move core directory to root * refactor and clean code
