--- app/core/Bootstrap.php
@@ -1,8 +1,0 @@
-<?php
-
-
-
-
-//require 'Request.php';
-
-//require 'Router.php';
