+++ composer.json
@@ -1,13 +1,8 @@
 {
-
   "autoload": {
-
     "psr-4": {
-
-      "App\\":"app/"
-
+      "App\\": "app/",
+      "Core\\": "core/"
     }
-
   }
-
 }
