+++ core/Bootstrap.php
@@ -1,0 +1,5 @@
+<?php
+ini_set('display_errors', 1);
+error_reporting(E_ALL);
+
+require 'vendor/autoload.php';
