+++ core/Request.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\core;
+namespace Core;
 /**
  * Request Class
  */
