+++ routes.php
@@ -1,4 +1,6 @@
 <?php
 
+/** @var Core\Router $router */
+
 $router->get('', 'SiteController@home');
 $router->get('about', 'SiteController@about');
